http://issues.apache.org/activemq/browse/AMQCPP-30
Fixes for Tight Marshalling.  

Changed the DataOutputStream writeBytes to not append a trailing null and made the writeChars do the trailing null append.  
Updated the docs on the methods to make this clear, and changed the code that was calling writeBytes to call writeChars 
in the places that a trailing null was expected.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@512268 13f79535-47bb-0310-9956-ffa450edef68