http://issues.apache.org/activemq/browse/AMQCPP-81
Changed the Connector interface so that when a consumer is created, two methods are called.  

createConsumer which sets up the consumer and returns a valid ConsumerInfo pointer.
startConsumer which actually registers the consumer with the broker.

This allows the consumer to be prepared but not actually started until the managing class is setup and ready for messages.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@518961 13f79535-47bb-0310-9956-ffa450edef68