http://issues.apache.org/activemq/browse/AMQCPP-104
Adjusted the TCPTransport to close its socket before calling the next close.  
Fixed the Socket Streams to have closed flags that are checked when the socket ops fail.
Changed the close method in TCPSocket to close the streams, but not delete them so that users don't get access violation errors when closing the socket with threads using the streams, streams are deleted in the dtor and in the connect method.



git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@534658 13f79535-47bb-0310-9956-ffa450edef68