http://issues.apache.org/activemq/browse/AMQCPP-108
MAde the connector use a Stomp Connection Negotiator that handles the connection phase, no other changes to the Connector other than to use this were made.  The negotiator requires that the ConnectedCommand is no longer a Response derived object, but once it gets it it creates a ConnectedWrapper instnace that is a response so that request can return it.  

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@541683 13f79535-47bb-0310-9956-ffa450edef68