https://issues.apache.org/activemq/browse/AMQCPP-195
Changed the way messages are handled in transacted sessions, now they are all ack'd even if the commit is done in the onMessage callback.  Also improved transaction support to not ack ever message until the comit is done so that one large ack can be done in openwire reducing network overhead.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@688552 13f79535-47bb-0310-9956-ffa450edef68