http://issues.apache.org/activemq/browse/AMQCPP-100
Refactored TaskRunner classes to add a composite task runner and a dedicated task runner for use failover transport.  The dedicated task runner will be used to close transports out of band to avoid deadlock when a transport is close as a result of a call to the onException method from the inner thread of the IOTransport.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@757843 13f79535-47bb-0310-9956-ffa450edef68