https://issues.apache.org/activemq/browse/AMQCPP-293
Move configuration options into the ActiveMQConnectionFactory and delay the send of ConnectionInfo until the Connection is used to allow a user to specify the clientId.  Adds setClientID to cms::Connection.  Add class IdGenerator which creates client and connection ids using the same method as the java and .net clients to make ids consistent across the clients.  Add getLocalHost to class InetAddress.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@951033 13f79535-47bb-0310-9956-ffa450edef68