https://issues.apache.org/activemq/browse/AMQCPP-316
Fix a case where the connection's config object could be leaked on a shutdown where the failover transport has its timeout option set and am exception is thrown to the connection when its RemoveInfo command or ShutdownInfo command is sent.

Also unset the exception listener eariler than it was before since we don't care about async errors once we enter the disconnect method.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@999121 13f79535-47bb-0310-9956-ffa450edef68