Move timer objects to the end of member list
In high CPU load it can happen, that Timer thread wakes up,
is swapped out, and when application resumes it does not resume with timer
thread, but transport thread, that decides to do failover, destroys
`InactivityMonitor` object. When its members are getting destroyed,
they arrive at `Timer` object that joins, timer thread resumes in
`writeCheck` or `readCheck` method, only to find out `asyncTasks` is already
destroyed because it comes later in member declaration list so it was
destroyed first. Moving `Timer` object to the end of the list, guarantees
that they will be destroyed first, so their threads will get joined and
allowed to finish, before the rest of the object is destroyed.

Closes AMQCPP-626